add prepare async payload function
